Initial Code
I am a work-study for student activities and one of the things that cause an issue is at the front desk... we get a lot of confusion and a lot of the same questions over and over again. This is my weak attempt at straining some of the commonly asked questions...probably much need for improvement but, hey it's a start.